Make debugging support slightly more generic
On windows, and init() function unconditionally looked for a file
with panic data, redirected stderr into a new file and copied the
panic data into that file.  This worked for editcp, but not for
dmrRadio.

We move that code into a named function which editcp calls on
startup.  On unix, that function is a no-op.